# docs/swagger.yaml
basePath: /
definitions:
...
  model.Chamado:
    properties:
      atribuidoId:
        type: string
      atualizadoEm:
        type: string
      categoriaId:
        type: string
      criadoEm:
        type: string
      criadorId:
        type: string
      descricao:
        type: string
      fechadoEm:
        type: string
      id:
        type: string
      solucao:
        type: string
      solucionadoEm:
        type: string
      status:
        $ref: '#/definitions/model.StatusChamado'
      subcategoriaId:
        type: string
      titulo:
        type: string
    type: object
...
  model.StatusChamado:
    enum:
    - ABERTO
    - ATRIBUIDO
    - RESOLVIDO
    - REJEITADO
    - FECHADO
    - ARQUIVADO
    type: string
    x-enum-varnames:
    - StatusAberto
    - StatusAtribuido
    - StatusResolvido
    - StatusRejeitado
    - StatusFechado
    - StatusArquivado
...
host: localhost:8080
info:
  contact: {}
  description: API para gestão de chamados
  title: Gestor de Chamados API
  version: "1.0"
paths:
...
  /chamados/arquivar/{id}:
    patch:
      consumes:
      - application/json
      description: Arquiva um chamado existente pelo ID.
      parameters:
      - description: ID do chamado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Arquiva um chamado existente
      tags:
      - chamados
  /chamados/atribuir-tecnico/{id}:
    patch:
      consumes:
      - application/json
      description: Atribui um técnico a um chamado existente pelo ID.
      parameters:
      - description: ID do chamado
        in: path
        name: id
        required: true
        type: string
      - description: ID do técnico
        in: body
        name: tecnico
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Atribui um técnico a um chamado existente
      tags:
      - chamados
  /chamados/atualizar-status/{id}:
    patch:
      consumes:
      - application/json
      description: Atualiza o status de um chamado existente pelo ID.
      parameters:
      - description: ID do chamado
        in: path
        name: id
        required: true
        type: string
      - description: Status e solução do chamado
        in: body
        name: chamado
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Atualiza o status de um chamado existente
      tags:
      - chamados
  /chamados/atualizar/{id}:
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um chamado existente pelo ID.
      parameters:
      - description: ID do chamado
        in: path
        name: id
        required: true
        type: string
      - description: Dados do chamado
        in: body
        name: chamado
        required: true
        schema:
          $ref: '#/definitions/model.Chamado'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Chamado'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Atualiza um chamado existente
      tags:
      - chamados
  /chamados/buscar-por-id/{id}:
    get:
      consumes:
      - application/json
      description: Retorna chamado pelo ID.
      parameters:
      - description: ID do chamado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Chamado'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Busca um chamado por ID
      tags:
      - chamados
  /chamados/buscar-tudo:
    get:
      consumes:
      - application/json
      description: Retorna lista paginada de chamados.
      parameters:
      - description: Pagina
        in: query
        name: pagina
        type: integer
      - description: Limite
        in: query
        name: limite
        type: integer
      - description: Busca
        in: query
        name: busca
        type: string
      - description: Status do chamado
        in: query
        name: status
        type: string
      - description: ID da categoria
        in: query
        name: categoriaId
        type: string
      - description: ID da subcategoria
        in: query
        name: subcategoriaId
        type: string
      - description: ID do criador
        in: query
        name: criadorId
        type: string
      - description: ID do atribuído
        in: query
        name: atribuidoId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Chamado'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Lista chamados com paginação e filtros
      tags:
      - chamados
  /chamados/criar:
    post:
      consumes:
      - application/json
      description: Cria um novo chamado com os dados fornecidos no corpo da requisição.
      parameters:
      - description: Dados do chamado
        in: body
        name: chamado
        required: true
        schema:
          $ref: '#/definitions/model.Chamado'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Chamado'
        "400":
          description: Bad Request
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Cria um novo chamado
      tags:
      - chamados
  /chamados/desarquivar/{id}:
    patch:
      consumes:
      - application/json
      description: Desarquiva um chamado existente pelo ID.
      parameters:
      - description: ID do chamado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Desarquiva um chamado existente
      tags:
      - chamados
  /chamados/lista-completa:
    get:
      consumes:
      - application/json
      description: Retorna todos os chamados sem paginação, útil para relatórios ou
        exportação de dados.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Chamado'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Retorna todos os chamados sem paginação
      tags:
      - chamados
  /chamados/remover-tecnico/{id}:
    delete:
      consumes:
      - application/json
      description: Remove o técnico atribuído de um chamado existente pelo ID.
      parameters:
      - description: ID do chamado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Remove o técnico atribuído de um chamado existente
      tags:
      - chamados
...
swagger: "2.0"